feat: bump version react 19, tailwind v4, complete non-vite source code
