+++ sources/b-end-non-vite/client/index.html
@@ -1,36 +1,38 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-    <!-- Use CDN for tailwind v3 -->
-    <script src="https://cdn.tailwindcss.com"></script>
+<head>
+  <meta charset="UTF-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-    <title>ToDo ReactJS v1</title>
-  </head>
+  <!-- Use CDN for tailwind v3 -->
+  <script src="https://cdn.jsdelivr.net/npm/@tailwindcss/browser@4"></script>
 
-  <body>
-    <!-- TODO - Langkah 3: Tambahkan sebuah root element untuk React -->
-    <div id="root"></div>
+  <title>ToDo ReactJS v1</title>
+</head>
 
-    <!-- TODO: Langkah 4: Pindahkan seluruh section ke dalam script React (JSX) -->
+<body>
+  <!-- TODO - Langkah 3: Tambahkan sebuah root element untuk React -->
+  <div id="root"></div>
 
-    <!-- Script section -->
-    <script src="//cdn.jsdelivr.net/npm/sweetalert2@11"></script>
+  <!-- TODO: Langkah 4: Pindahkan seluruh section ke dalam script React (JSX) -->
 
-    <!-- Reference: https://gist.githubusercontent.com/gaearon/0275b1e1518599bbeafcde4722e79ed1/raw/db72dcbf3384ee1708c4a07d3be79860db04bff0/example.html -->
+  <!-- Script section -->
+  <script src="//cdn.jsdelivr.net/npm/sweetalert2@11"></script>
 
-    <!-- TODO - Langkah 1: Tambahkan script untuk menggunakan React -->
-    <script src="https://unpkg.com/react@18/umd/react.development.js"></script>
-    <script src="https://unpkg.com/react-dom@18/umd/react-dom.development.js"></script>
+  <!-- Reference: https://gist.githubusercontent.com/gaearon/0275b1e1518599bbeafcde4722e79ed1/raw/db72dcbf3384ee1708c4a07d3be79860db04bff0/example.html -->
 
-    <!-- TODO - Langkah 2: Tambahkan script untuk mengkonversi React menjadi sesuatu yang bisa dibaca oleh browser (babel) -->
-    <!-- JANGAN PERNAH MENGGUNAKAN INI DI PRODUCTION ! -->
-    <script src="https://unpkg.com/@babel/standalone/babel.min.js"></script>
+  <!-- TODO - Langkah 1: Tambahkan script untuk menggunakan React -->
+  <!-- <script src="https://unpkg.com/react@18/umd/react.development.js"></script>
+  <script src="https://unpkg.com/react-dom@18/umd/react-dom.development.js"></script> -->
 
-    <!-- TODO - Langkah 5: Gunakan React -->
-    <script type="text/babel" src="./src/main.js"></script>
-  </body>
+  <!-- TODO - Langkah 2: Tambahkan script untuk mengkonversi React menjadi sesuatu yang bisa dibaca oleh browser (babel) -->
+  <!-- JANGAN PERNAH MENGGUNAKAN INI DI PRODUCTION ! -->
+  <script src="https://unpkg.com/@babel/standalone/babel.min.js"></script>
+
+  <!-- TODO - Langkah 5: Gunakan React -->
+  <script type="text/babel" data-type="module" src="./src/main.jsx"></script>
+</body>
+
 </html>
