+++ sources/b-end-with-vite/client/index.html
@@ -1,17 +1,20 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <!-- TODO - Langkah 1: Ganti Title dan Tambahkan script untuk CSS yang dibutuhkan -->
-    <title>ToDo ReactJS v1</title>
 
-    <!-- Use CDN for tailwind v3 -->
-    <script src="https://cdn.tailwindcss.com"></script>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
-  </body>
+<head>
+  <meta charset="UTF-8" />
+  <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <!-- TODO - Langkah 1: Ganti Title dan Tambahkan script untuk CSS yang dibutuhkan -->
+  <title>ToDo ReactJS v1</title>
+
+  <!-- Use CDN for tailwind v3 -->
+  <script src="https://cdn.jsdelivr.net/npm/@tailwindcss/browser@4"></script>
+</head>
+
+<body>
+  <div id="root"></div>
+  <script type="module" src="/src/main.jsx"></script>
+</body>
+
 </html>
